Various small updates
Mostly just reviewed the current status. My opinion is that there needs
to be a way to handle the separate enemy shots. Can't let each enemy
shot type have it's own list.

The other thing is that "World" is becoming a "God object" and is too
big. This is mostly due to the sheer number of separate lists it's
managing.

Again, since the design philosophy seems to be "have a separate list for
each type of object", we need to either fix that design philosophy or
streamline it somehow.